update user functionality added , stored all user in one localstorage object and fixing and improvment

+++ Evaluation/JavaScript/UserScript.js
@@ -1,42 +1,115 @@
-let loggedUser = JSON.parse(localStorage.getItem("AdminInfo"));
-
-if (loggedUser.isAdmin) {
-  $("#loginUserName").html(`Hello ,${loggedUser.name}`);
-} else {
-  // console.log();
-}
-
-displayUser();
-
+let loggedUser = JSON.parse(localStorage.getItem("loggedUser"));
+$("#loginUserName").html(`Hello ,${loggedUser.name}`);
+var isUserUpdate = false;
+RenderUsers();
+//add user
 $("#addUserForm").submit(function (e) {
   e.preventDefault();
+
   let name = $("#name").val();
   let email = $("#email").val();
   let password = $("#password").val();
   let birthdate = $("#birthdate").val();
   const formValid = formValidation(name, email, password, birthdate);
+
   if (formValid) {
-    const usersData = JSON.parse(localStorage.getItem("usersData")) || [];
-    let user_id = Date.now();
-    //isAdmin is false that means this all user are not admin
-    let user = {
-      id: user_id,
-      name: name,
-      email: email,
-      password: password,
-      birthdate: birthdate,
-      isAdmin: false,
-    };
-    usersData.push(user);
-    localStorage.setItem("usersData", JSON.stringify(usersData));
-    addUserForm.reset();
-    alert("user added");
-    displayUser();
+    let tableData = JSON.parse(localStorage.getItem("usersData")) || [];
+    let updateUserId = parseInt(
+      document.getElementById("AddUserButton").getAttribute("data-id")
+    );
+    if (updateUserId) {
+      console.log(typeof updateUserId);
+      updateUser(updateUserId);
+    } else {
+      console.log("If else working, in else section");
+      let user_id = Date.now();
+      tableData = [
+        ...tableData,
+        {
+          id: user_id,
+          name: name,
+          email: email,
+          password: password,
+          birthdate: birthdate,
+          isAdmin: false,
+        },
+      ];
+      localStorage.setItem("usersData", JSON.stringify(tableData));
+      addUserForm.reset();
+      alert("New user Added");
+      RenderUsers();
+    }
   }
 });
 
-//display user function
-function displayUser() {
+function calculateAge(birthYear) {
+  console.log(birthYear);
+}
+function BindEditUserDate(userId) {
+  isUserUpdate = true;
+  console.log("bind user data");
+  console.log(typeof userId);
+  $("#AddUserButton").attr("data-id", userId); //adding data-id so we can check form is for updating user or adding new user
+
+  $("#AddUserButton").attr("value", "Update User");
+  $("#userPageHeading").html("Update User");
+  let tableData = JSON.parse(localStorage.getItem("usersData")) || [];
+  let userDetails = tableData.find((user) => user.id === userId);
+  // bind user data to input felid
+  $("#name").val(userDetails.name);
+  $("#email").val(userDetails.email);
+  $("#password").val(userDetails.password);
+  $("#birthdate").val(userDetails.birthdate);
+}
+function updateUser(userId) {
+  let name = $("#name").val();
+  let email = $("#email").val();
+  let password = $("#password").val();
+  let birthdate = $("#birthdate").val();
+  //get the value of uesrinput feild
+
+  let tableData = JSON.parse(localStorage.getItem("usersData")) || [];
+  if (tableData && tableData.length > 0) {
+    const indexOfuser = tableData.findIndex((user) => user.id === userId);
+    console.log(tableData[indexOfuser]);
+    if (indexOfuser >= 0) {
+      //editing the user
+      tableData[indexOfuser] = {
+        ...tableData[indexOfuser],
+        name: name,
+        email: email,
+        password: password,
+        birthdate: birthdate,
+      };
+      console.log(tableData[indexOfuser]);
+      localStorage.setItem("usersData", JSON.stringify(tableData));
+    }
+    RenderUsers();
+  }
+  $("#addUserForm")[0].reset();
+  alert("User Updated");
+  $("#AddUserButton").attr("value", "Add User");
+  $("#userPageHeading").html("Add User");
+}
+
+//delete user function
+function deleteUser(userId) {
+  let confirmed = confirm("Are you sure you want to delete this user?");
+  if (confirmed) {
+    let tableData = JSON.parse(localStorage.getItem("usersData")) || [];
+
+    let userIndex = tableData.findIndex((user) => user.id === userId);
+    if (userIndex >= 0) {
+      tableData.splice(userIndex, 1);
+      localStorage.setItem("usersData", JSON.stringify(tableData));
+    }
+    RenderUsers();
+  }
+}
+
+//RederUser function
+function RenderUsers() {
+  console.log("new redner");
   let tbody = document.getElementsByTagName("tbody")[0];
   tbody.innerHTML = "";
   const usersData = JSON.parse(localStorage.getItem("usersData")) || [];
@@ -49,45 +122,27 @@ function displayUser() {
     row.appendChild(td);
     tbody.appendChild(row);
     return;
-  }
-  usersData.forEach((user) => {
-    const row = document.createElement("tr");
-    row.innerHTML = `
-        <td>${user.name}</td>
-        <td>${user.email}</td>
-        <td>${user.password}</td>
-        <td>${user.birthdate}</td>
-        <td></td>
-
-        <td>
-          <button onClick="editUser(${user});">Edit</button>
-          <button onClick="deleteUser(${user.id});">Delete</button>
-        </td>
-      `;
-    tbody.appendChild(row);
-  });
-}
-function calculateAge(birthYear) {
-  console.log(birthYear);
-}
-function editUser(user) {
-  $("#name").val(user.name);
-  $("#email").val(user.email);
-  $("#password").val(user.password);
-  $("#birthdate").val(user.birthdate);
-}
-function deleteUser(userId) {
-  let conform = confirm("Are you sure to delete this user?");
-  if (conform) {
-    const usersData = JSON.parse(localStorage.getItem("usersData")) || [];
-    const newusersData = usersData.filter((user) => {
-      return userId != user.id;
+  } else {
+    usersData.forEach((user) => {
+      //only display sub user details
+      if (!user.isAdmin) {
+        const row = document.createElement("tr");
+        row.innerHTML = `
+            <td>${user.name}</td>
+            <td>${user.email}</td>
+            <td>${user.password}</td>
+            <td>${user.birthdate}</td>
+            <td></td>
+            <td>
+              <button onClick="BindEditUserDate(${user.id});">Edit</button>
+              <button onClick="deleteUser(${user.id});">Delete</button>
+            </td>
+          `;
+        tbody.appendChild(row);
+      }
     });
-    localStorage.setItem("usersData", JSON.stringify(newusersData));
-    displayUser();
   }
 }
-
 //form validation function
 function formValidation(name, email, password, birthdate) {
   let valid = true;
